requirements.txt
django
djangorestframework
djangorestframework-simplejwt
stripe
python-decouple